[DO NOT MERGE]: truncate range proofs
This truncates rangeproofs in serialization/deserialization, and
enables sentinels to recover the appropriate commitment/rangeproof
combination to ignore what was sent over the wire and undergo its
standard verification routine.

In addition, the size of rangeproofs in compact_outputs are modified
to reflect the truncated size (such that shards only store the
truncated length).

This should have the effect of a reasonable mock for how the system
will behave if we implemented the inner-product argument optimization
from bulletproofs (both reduction in bandwidth and storage
requirements).

Signed-off-by: Sam Stuewe <stuewe@mit.edu>